+++ crates/miden-lib/src/account/components/mod.rs
@@ -26,6 +26,16 @@ static BASIC_FUNGIBLE_FAUCET_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
     Library::read_from_bytes(bytes).expect("Shipped Basic Fungible Faucet library is well-formed")
 });
 
+// Initialize the Rpo Falcon 512 Procedure ACL library only once.
+static RPO_FALCON_512_PROCEDURE_ACL_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
+    let bytes = include_bytes!(concat!(
+        env!("OUT_DIR"),
+        "/assets/account_components/rpo_falcon_512_procedure_acl.masl"
+    ));
+    Library::read_from_bytes(bytes)
+        .expect("Shipped Rpo Falcon 512 Procedure ACL library is well-formed")
+});
+
 /// Returns the Basic Wallet Library.
 pub fn basic_wallet_library() -> Library {
     BASIC_WALLET_LIBRARY.clone()
@@ -39,4 +49,9 @@ pub fn rpo_falcon_512_library() -> Library {
 /// Returns the Basic Fungible Faucet Library.
 pub fn basic_fungible_faucet_library() -> Library {
     BASIC_FUNGIBLE_FAUCET_LIBRARY.clone()
+}
+
+/// Returns the Rpo Falcon 512 Procedure ACL Library.
+pub fn rpo_falcon_512_procedure_acl_library() -> Library {
+    RPO_FALCON_512_PROCEDURE_ACL_LIBRARY.clone()
 }